requirements.txt
tensorflow-cpu>=2.4.0
opencv-python-headless>=4.5.0
streamlit>=1.0.0
numpy>=1.19.0
pillow>=8.0.0
matplotlib>=3.3.0
streamlit-webrtc>=0.45.0
protobuf>=3.20.0,<4.0.0 